Move parsing function. Fix logging message
